Update List of Azure Resource Provider Name Space --> Azure Service
I have a large customer that has reached out to me letting me know that our MSFT documentation is out of date. Apparently, the list of Resource Provider namespaces does not align to what they see within their Azure subscription. I have also confirmed this with my subscription as well. An example of this is AKS (Microsoft.Kubernetes) for example. Suggest this be updated since customers are referencing it.